Test stdout and stderr specifically in Cucumber scenarios
Currently, keep_up does not capture stderr from calls to bundler, so any
errors that bundler produces can be examined by the user. This means
that extra output from bundler to stderr can cause Cucumber scenarios to
start failing if we test the whole output.

Since we are only really interested in testing the output from keep_up
itself, and stdout from bundler is processed internally and never output
to the user, it is better to test the stdout specifically, except where
keep_up outputs to stderr.