Speed up preparation of remote gem specifications in cucumber scenarios
This groups the specification of remote gems together into a single
step, and generates the gem index only once. Since generating the index
is quite slow, this saves quite some time per scenario.

The downside is that the contents of the lockfile need to be specified
explicitly, rather than relying on bundle install to create it correctly
for the given scenario. Given the resulting speed-up, I think this is
worth it.